# Conflicts:
#	paypro-android-backend/src/main/java/air/found/payproandroidbackend/business_logic/MerchantService.java
#	paypro-android-backend/src/main/resources/application.properties

+++ paypro-android-backend/src/main/java/air/found/payproandroidbackend/business_logic/MerchantService.java
@@ -1,11 +1,11 @@
 package air.found.payproandroidbackend.business_logic;
 
-import air.found.payproandroidbackend.core.models.Merchants;
+import air.found.payproandroidbackend.core.models.Merchant;
 import air.found.payproandroidbackend.data_access.MerchantRepository;
-import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 public class MerchantService {
@@ -17,23 +17,13 @@ public class MerchantService {
     }
 
     public boolean deleteMerchantById(Integer merchantId) {
-        try {
-            merchantsRepository.deleteById(merchantId);
-            return true; // Deletion successful
-        } catch (EmptyResultDataAccessException e) {
-            return false; // Merchant not found
-        } catch (Exception e) {
-            return false; // Error deleting merchant
-        }
-    }
+        Optional<Merchant> merchantOptional = merchantsRepository.findById(merchantId);
 
-    @Transactional
-    public Merchants addMerchant(Merchants merchant) {
-        try {
-            return merchantsRepository.save(merchant);
-        }
-        catch (Exception ex) {
-            return null;
+        if (merchantOptional.isPresent()) {
+            merchantsRepository.deleteById(merchantId);
+            return true;
+        } else {
+            return false;
         }
     }
 }
